requirements.txt
tqdm
numpy
pandas
plotly
openai
streamlit
faiss-cpu
python-dotenv
sentence-transformers